shutil move test & junit two environment(pass/failure)

+++ Src_Module/test/Verification_HumanEval_Test.py
@@ -1,6 +1,8 @@
+import os
 import re
 import unittest
 
+from Config import ROOT
 from Src_Module.src.Verification import Verification
 from Src_Module.src.Verification_HumanEval import Verification_HumanEval
 from Util_Module.src.FileIO import FileIO
@@ -51,28 +53,42 @@ class Verification_HumanEval_Test(unittest.TestCase):
 
     def test_write_data_in_junit_environment(self):
         self.verification_HumanEval.setJunitEnvironment('JUnit_Environment/HumanEval')
-        self.verification_HumanEval.setGoogleJavaFormat('Tool/google-java-format-1.18.1-all-deps.jar')
 
         javaCode = f"""
-        public class 'ADD_TEST_1' {{
+        public class ADD_TEST_1 {{
         public static int add(int x,int y){{
             return x + y;
         }}
         }}
         """
 
-        self.verification_HumanEval.junitEnvironment_Initialize()
-        self.assertEqual(len(self.fileIO.getFileListUnderFolder(self.verification_HumanEval.getJunitEnvironment())), 0)
+        self.verification_HumanEval.junitEnvironment_Initialize(self.verification_HumanEval.getJunitEnvironmentPass())
+        self.assertEqual(len(self.fileIO.getFileListUnderFolder(self.verification_HumanEval.getJunitEnvironmentPass())), 0)
 
-        target = self.verification_HumanEval.getJunitEnvironment() + '/Module_{}/{}.java'.format('Add', 'Add_Test_1')
+        self.verification_HumanEval.junitEnvironment_Initialize(self.verification_HumanEval.getJunitEnvironmentFailure())
+        self.assertEqual(len(self.fileIO.getFileListUnderFolder(self.verification_HumanEval.getJunitEnvironmentFailure())), 0)
+
+
+        target = self.verification_HumanEval.getJunitEnvironmentFailure() + '/Module_{}/{}.java'.format('Add', 'Add_TEST_1')
+        if self.fileIO.isPathExist(target):
+            self.fileIO.deleteFileData(target)
+
         self.fileIO.writeFileData(target, javaCode)
         self.assertTrue(self.fileIO.isPathExist(target))
 
         readData = self.fileIO.readFileData(target)
         self.assertEqual(self.normalize(readData), self.normalize(javaCode))
 
-        self.fileIO.deleteFileData(target)
-        self.assertFalse(self.fileIO.isPathExist(target))
+    def test_check_java_format_Pass(self):
+        self.verification_HumanEval.setJunitEnvironment('JUnit_Environment/HumanEval')
+        self.verification_HumanEval.setGoogleJavaFormat('Tool/google-java-format-1.18.1-all-deps.jar')
+        target = self.verification_HumanEval.getJunitEnvironmentFailure() + '/Module_{}/{}.java'.format('Add', 'Add_TEST_1')
+        result = self.verification_HumanEval.subprocess_run_JavaFormat(target)
+
+        self.assertEqual(result.returncode, 0)
+
+
+
 
 
 if __name__ == '__main__':
